Update reflected radiation model
Add surface albedo multipler to the denominator of the calculation that
computes potential surface-reflected outgoing solar radiation. This follows
Salamanca et al. (2012). Physically, this builds in an assumption that sunlight
reflected from a brightened surface and then reflected back down by the atmosphere
is re-reflected according to the albedo of the surrounding area. We use the surface
albedo computed from the radiative fluxes in ERA5 for the new calculation.