Merge `feature/nearby-albedo` into `main` (#16)
* Update unreleased changes

* Update reflected radiation model

Add surface albedo multiplier to the denominator of the calculation that
computes potential surface-reflected outgoing solar radiation. This follows
Salamanca et al. (2012). Physically, this builds in an assumption that sunlight
reflected from a brightened surface and then reflected back down by the atmosphere
is re-reflected according to the albedo of the surrounding area. We use the surface
albedo computed from the radiative fluxes in ERA5 for the new calculation.

* Update notebooks for new cloud storage settings

The Google Cloud Storage buckets with data for this project have been
changed to requester pays. This means that the functions that interact
with Google Cloud Storage must specify the project id so that billing
is handled appropriately.

* Fix typo in mamba/conda command

* Update data assets

* Update notebook for change in cloud storage settings

* Update Figure 3 for change in reflected radiation model

* Update ROM image version

* Update CHANGELOG